requirements.txt
Jinja==1.2
PyYAML==3.11
termcolor==1.1.0
